meme-api/requirements.txt
# FastAPI and server dependencies
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-dotenv>=1.0.0
slowapi>=0.1.9

# LangChain and LangGraph dependencies
langgraph>=1.0.0
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-openai>=1.1.0
langchain-google-genai>=4.2.0

# Google AI dependencies
google-genai>=1.58.0
google-generativeai>=0.8.0

# OpenAI dependencies
openai>=1.12.0

# Image processing (for image analysis only)
Pillow>=12.0.0

# Utilities
aiofiles>=23.2.0
